requirements.txt
anyio==3.6.2
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.3.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.6
Django==4.1.7
django-ckeditor==6.5.1
django-cors-headers==4.2.0
django-js-asset==2.0.0
djangorestframework==3.14.0
dnspython==2.3.0
email-validator==1.3.1
executing==1.2.0
fastapi==0.95.0
filelock==3.9.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
ipykernel==6.21.1
ipython==8.9.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.2
jupyter_core==5.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
orjson==3.8.9
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pipenv==2023.2.4
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
pydantic==1.10.7
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
pywin32==305
PyYAML==6.0
pyzmq==25.0.0
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.26.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
ujson==5.7.0
uritemplate==4.1.1
urllib3==2.0.5
uvicorn==0.21.1
virtualenv==20.18.0
virtualenv-clone==0.5.7
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.1
